[⚙]🐛fix: 회원 탈퇴 데이터 반영

+++ src/main/java/ddwu/project/mdm_ver2/domain/product/entity/Product.java
@@ -38,7 +38,7 @@ public class Product implements Serializable {
     private String imgUrl;
 
     @Column(name = "review_cnt")
-    private int reviewCnt;
+    private Long reviewCnt;
 
     @Column(name = "review_star_avg")
     private float reviewStarAvg;
